refactor: break specs↔spec-viewer layer violation (Phase 9)
specExplorerProvider (sidebar) was importing isStepCompleted from
spec-viewer/stateDerivation and fileNameToDisplayName from
spec-viewer/utils. Sidebar shouldn't depend on viewer — wrong direction.

Moves:
  - isStepCompleted → src/features/specs/stepHistoryDerivation.ts
    (it's a query over StepName + stepHistory, belongs in the canonical
    specs layer alongside deriveStepHistory)
  - fileNameToDisplayName + fileNameToDocType → src/core/utils/fileNaming.ts
    (pure label helpers, neither sidebar nor viewer owns them)

The old locations now re-export from the new ones so existing test
suites and the viewer's internal callers keep working unchanged. The
re-exports are deletable on a future pass once external imports migrate.

Import direction is now one-way: spec-viewer → specs → core. Sidebar
doesn't reach into the viewer module anymore.

All 729 tests pass.
